+++ SerenityLib/examples/solarSystem/game/weapons/PhotonTorpedoOld.h
@@ -66,7 +66,7 @@ class PhotonTorpedoOld final : public SecondaryWeaponTorpedo {
             const glm::vec3& forward,
             const float& arc,
             const uint& maxCharges = 1,
-            const uint& damage = 1650,
+            const float& damage = 1650.0f,
             const float& _rechargePerRound = 4.5f,
             const float& _impactRadius = 5.5f,
             const float& _impactTime = 3.1f,
